+++ packages/stream_chat_flutter/lib/src/attachment/url_attachment.dart
@@ -37,11 +37,11 @@ class UrlAttachment extends StatelessWidget {
     final chatThemeData = StreamChatTheme.of(context);
     return GestureDetector(
       onTap: () {
-        final titleLink = urlAttachment.titleLink;
-        if (titleLink != null) {
+        final ogScrapeUrl = urlAttachment.ogScrapeUrl;
+        if (ogScrapeUrl != null) {
           onLinkTap != null
-              ? onLinkTap!(titleLink)
-              : launchURL(context, titleLink);
+              ? onLinkTap!(ogScrapeUrl)
+              : launchURL(context, ogScrapeUrl);
         }
       },
       child: Column(