+++ packages/stream_chat_flutter/lib/src/channel_header.dart
@@ -61,9 +61,11 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
     this.showConnectionStateTile = false,
     this.title,
     this.subtitle,
+    this.centerTitle,
     this.leading,
     this.actions,
     this.backgroundColor,
+    this.elevation = 1,
   })  : preferredSize = const Size.fromHeight(kToolbarHeight),
         super(key: key);
 
@@ -92,6 +94,9 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
   /// Subtitle widget
   final Widget? subtitle;
 
+  /// Whether the title should be centered
+  final bool? centerTitle;
+
   /// Leading widget
   final Widget? leading;
 
@@ -102,8 +107,16 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
   /// The background color for this [ChannelHeader].
   final Color? backgroundColor;
 
+  /// The elevation for this [ChannelHeader].
+  final double elevation;
+
   @override
   Widget build(BuildContext context) {
+    final effectiveCenterTitle = getEffectiveCenterTitle(
+      Theme.of(context),
+      actions: actions,
+      centerTitle: centerTitle,
+    );
     final channel = StreamChannel.of(context).channel;
     final channelHeaderTheme = ChannelHeaderTheme.of(context);
 
@@ -144,7 +157,7 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
             systemOverlayStyle: theme.brightness == Brightness.dark
                 ? SystemUiOverlayStyle.light
                 : SystemUiOverlayStyle.dark,
-            elevation: 1,
+            elevation: elevation,
             leading: leadingWidget,
             backgroundColor: backgroundColor ?? channelHeaderTheme.color,
             actions: actions ??
@@ -162,14 +175,16 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
                     ),
                   ),
                 ],
-            centerTitle: true,
+            centerTitle: centerTitle,
             title: InkWell(
               onTap: onTitleTap,
               child: SizedBox(
                 height: preferredSize.height,
-                width: preferredSize.width,
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: effectiveCenterTitle
+                      ? CrossAxisAlignment.center
+                      : CrossAxisAlignment.stretch,
                   children: <Widget>[
                     title ??
                         ChannelName(