+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations.dart
@@ -74,6 +74,8 @@ GlobalStreamChatLocalizations? getStreamChatTranslation(Locale locale) {
       return const StreamChatLocalizationsKo();
     case 'pt':
       return const StreamChatLocalizationsPt();
+    default:
+      return null;
   }
 }